pyproject.toml
[project]
name = "langchain-parsers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.2",
    "langchain-google-genai>=3.2.0",
    "python-dotenv>=1.2.1",
]
